[Refactor] add file at gitignore and rename directory domain
