+++ src/main/java/com/capstone/favicon/dataset/domain/FileExtension.java
@@ -1,4 +1,4 @@
-package com.capstone.favicon.dataset.domain.domain;
+package com.capstone.favicon.dataset.domain;
 
 public enum FileExtension {
