+++ src/main/java/com/capstone/favicon/dataset/repository/DatasetRepository.java
@@ -1,6 +1,6 @@
 package com.capstone.favicon.dataset.repository;
 
-import com.capstone.favicon.dataset.domain.domain.Dataset;
+import com.capstone.favicon.dataset.domain.Dataset;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
