+++ src/main/java/com/capstone/favicon/dataset/repository/ResourceRepository.java
@@ -1,6 +1,6 @@
 package com.capstone.favicon.dataset.repository;
 
-import com.capstone.favicon.dataset.domain.domain.Resource;
+import com.capstone.favicon.dataset.domain.Resource;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
