+++ src/main/java/com/capstone/favicon/download/controller/s3FileDownloadController.java
@@ -12,8 +12,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 
 @RestController
 @RequestMapping("/data-set")
